+++ src/main/kotlin/xyz/xenondevs/nova/tileentity/impl/agriculture/Fertilizer.kt
@@ -48,9 +48,9 @@ import xyz.xenondevs.nova.world.region.Region
 import xyz.xenondevs.nova.world.region.VisualRegion
 import java.util.*
 
-private val MAX_ENERGY = NovaConfig[FERTILIZER].getInt("capacity")!!
-private val ENERGY_PER_TICK = NovaConfig[FERTILIZER].getInt("energy_per_tick")!!
-private val ENERGY_PER_FERTILIZE = NovaConfig[FERTILIZER].getInt("energy_per_fertilize")!!
+private val MAX_ENERGY = NovaConfig[FERTILIZER].getLong("capacity")!!
+private val ENERGY_PER_TICK = NovaConfig[FERTILIZER].getLong("energy_per_tick")!!
+private val ENERGY_PER_FERTILIZE = NovaConfig[FERTILIZER].getLong("energy_per_fertilize")!!
 private val IDLE_TIME = NovaConfig[FERTILIZER].getInt("idle_time")!!
 private val MIN_RANGE = NovaConfig[FERTILIZER].getInt("range.min")!!
 private val MAX_RANGE = NovaConfig[FERTILIZER].getInt("range.max")!!
