+++ src/main/kotlin/xyz/xenondevs/nova/tileentity/impl/agriculture/TreeFactory.kt
@@ -51,8 +51,8 @@ private val PLANTS = mapOf(
     Material.BROWN_MUSHROOM to PlantConfiguration(NovaMaterialRegistry.GIANT_BROWN_MUSHROOM_MINIATURE, ItemStack(Material.BROWN_MUSHROOM, 3), Color(149, 112, 80))
 )
 
-private val MAX_ENERGY = NovaConfig[TREE_FACTORY].getInt("capacity")!!
-private val ENERGY_PER_TICK = NovaConfig[TREE_FACTORY].getInt("energy_per_tick")!!
+private val MAX_ENERGY = NovaConfig[TREE_FACTORY].getLong("capacity")!!
+private val ENERGY_PER_TICK = NovaConfig[TREE_FACTORY].getLong("energy_per_tick")!!
 private val PROGRESS_PER_TICK = NovaConfig[TREE_FACTORY].getDouble("progress_per_tick")!!
 private val IDLE_TIME = NovaConfig[TREE_FACTORY].getInt("idle_time")!!
