+++ src/main/kotlin/xyz/xenondevs/nova/tileentity/impl/processing/ElectricalFurnace.kt
@@ -47,8 +47,8 @@ private val RECIPES: List<FurnaceRecipe> by lazy {
 private fun getRecipe(input: ItemStack) =
     RECIPES.firstOrNull { it.input.isSimilar(input) || it.inputChoice.test(input) }
 
-private val MAX_ENERGY = NovaConfig[ELECTRICAL_FURNACE].getInt("capacity")!!
-private val ENERGY_PER_TICK = NovaConfig[ELECTRICAL_FURNACE].getInt("energy_per_tick")!!
+private val MAX_ENERGY = NovaConfig[ELECTRICAL_FURNACE].getLong("capacity")!!
+private val ENERGY_PER_TICK = NovaConfig[ELECTRICAL_FURNACE].getLong("energy_per_tick")!!
 private val COOK_SPEED = NovaConfig[ELECTRICAL_FURNACE].getInt("cook_speed")!!
 
 class ElectricalFurnace(
