+++ src/main/kotlin/xyz/xenondevs/nova/tileentity/network/energy/EnergyBridge.kt
@@ -9,6 +9,6 @@ interface EnergyBridge : NetworkBridge {
      * types of [NetworkBridge]s inside an [EnergyNetwork], the transfer of the
      * whole network is equal to the smallest one.
      */
-    val energyTransferRate: Int
+    val energyTransferRate: Long
     
 }
