+++ src/main/kotlin/xyz/xenondevs/nova/tileentity/network/energy/holder/BufferEnergyHolder.kt
@@ -6,16 +6,16 @@ import xyz.xenondevs.nova.tileentity.network.energy.EnergyConnectionType
 
 class BufferEnergyHolder(
     endPoint: NetworkedTileEntity,
-    defaultMaxEnergy: Int,
+    defaultMaxEnergy: Long,
     private val creative: Boolean,
     lazyDefaultConfig: () -> MutableMap<BlockFace, EnergyConnectionType>
 ) : EnergyHolder(endPoint, defaultMaxEnergy, null, lazyDefaultConfig) {
     
-    override val requestedEnergy: Int
-        get() = if (creative) Int.MAX_VALUE else maxEnergy - energy
+    override val requestedEnergy: Long
+        get() = if (creative) Long.MAX_VALUE else maxEnergy - energy
     
-    override var energy: Int
-        get() = if (creative) Int.MAX_VALUE else super.energy
+    override var energy: Long
+        get() = if (creative) Long.MAX_VALUE else super.energy
         set(value) {
             if (creative) return
             super.energy = value
